Moved tests to test_hitl.py
